refactor: Типизация news.ts
